Move detail page factories to dedicated module
This maps closer to the model stucture.
