+++ network-api/networkapi/wagtailpages/factory/research_hub/__init__.py
@@ -1,13 +1,11 @@
-import random
-
 import factory
 import wagtail_factories
 
 from networkapi.utility.faker import helpers as faker_helpers
 from networkapi.wagtailpages import models as wagtailpage_models
-from networkapi.wagtailpages.factory import documents as documents_factory
 from networkapi.wagtailpages.factory import image_factory
 from networkapi.wagtailpages.factory import profiles as profiles_factory
+from networkapi.wagtailpages.factory.research_hub import detail_page as detail_page_factory
 
 
 class ResearchLandingPageFactory(wagtail_factories.PageFactory):
@@ -35,68 +33,6 @@ class ResearchAuthorsIndexPageFactory(wagtail_factories.PageFactory):
     banner_image = factory.SubFactory(image_factory.ImageFactory)
 
 
-class ResearchDetailLinkFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = wagtailpage_models.ResearchDetailLink
-
-    label = factory.Faker("text", max_nb_chars=30)
-    url = ""
-    document = None
-
-    class Params:
-        with_url = factory.Trait(
-            url=factory.Faker("uri"),
-        )
-        with_document = factory.Trait(document=factory.SubFactory(documents_factory.DocumentFactory))
-
-
-class ResearchDetailPageFactory(wagtail_factories.PageFactory):
-    class Meta:
-        model = wagtailpage_models.ResearchDetailPage
-
-    title = factory.Faker("text", max_nb_chars=50)
-    cover_image = factory.SubFactory(image_factory.ImageFactory)
-    research_links = factory.RelatedFactoryList(
-        factory=ResearchDetailLinkFactory,
-        factory_related_name="research_detail_page",
-        size=lambda: random.randint(1, 2),
-        with_url=True,
-    )
-    original_publication_date = factory.Faker("date_object")
-    introduction = factory.Faker("text", max_nb_chars=300)
-
-    @factory.lazy_attribute
-    def overview(self):
-        faker = faker_helpers.get_faker()
-        return "\n\n".join(faker.paragraphs(nb=3))
-
-    @factory.lazy_attribute
-    def collaborators(self):
-        faker = faker_helpers.get_faker()
-        names = []
-        for _ in range(random.randint(1, 5)):
-            names.append(faker.name())
-        return "; ".join(names)
-
-    research_authors = factory.RelatedFactoryList(
-        factory="networkapi.wagtailpages.factory.research_hub.ResearchAuthorRelationFactory",
-        factory_related_name="research_detail_page",
-        size=1,
-    )
-
-    related_topics = factory.RelatedFactoryList(
-        factory="networkapi.wagtailpages.factory.research_hub.ResearchDetailPageResearchTopicRelationFactory",
-        factory_related_name="research_detail_page",
-        size=1,
-    )
-
-    related_regions = factory.RelatedFactoryList(
-        factory="networkapi.wagtailpages.factory.research_hub.ResearchDetailPageResearchRegionRelationFactory",
-        factory_related_name="research_detail_page",
-        size=1,
-    )
-
-
 class ResearchRegionFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = wagtailpage_models.ResearchRegion
@@ -116,7 +52,7 @@ class ResearchAuthorRelationFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = wagtailpage_models.ResearchAuthorRelation
 
-    research_detail_page = factory.SubFactory(ResearchDetailPageFactory)
+    research_detail_page = factory.SubFactory(detail_page_factory.ResearchDetailPageFactory)
     author_profile = factory.SubFactory(profiles_factory.ProfileFactory)
 
 
@@ -124,7 +60,7 @@ class ResearchDetailPageResearchRegionRelationFactory(factory.django.DjangoModel
     class Meta:
         model = wagtailpage_models.ResearchDetailPageResearchRegionRelation
 
-    research_detail_page = factory.SubFactory(ResearchDetailPageFactory)
+    research_detail_page = factory.SubFactory(detail_page_factory.ResearchDetailPageFactory)
     research_region = factory.SubFactory(ResearchRegionFactory)
 
 
@@ -132,7 +68,7 @@ class ResearchDetailPageResearchTopicRelationFactory(factory.django.DjangoModelF
     class Meta:
         model = wagtailpage_models.ResearchDetailPageResearchTopicRelation
 
-    research_detail_page = factory.SubFactory(ResearchDetailPageFactory)
+    research_detail_page = factory.SubFactory(detail_page_factory.ResearchDetailPageFactory)
     research_topic = factory.SubFactory(ResearchTopicFactory)
 
 
@@ -170,7 +106,7 @@ def generate(seed):
         ResearchTopicFactory.create()
 
     for _ in range(13):
-        research_detail_page = ResearchDetailPageFactory.create(
+        research_detail_page = detail_page_factory.ResearchDetailPageFactory.create(
             parent=research_library_page,
             research_authors=None,
             related_topics=None,
