+++ network-api/networkapi/wagtailpages/tests/research_hub/test_author_index.py
@@ -6,6 +6,7 @@ from wagtail_localize import synctree
 
 from networkapi.wagtailpages.factory import profiles as profiles_factory
 from networkapi.wagtailpages.factory import research_hub as research_factory
+from networkapi.wagtailpages.factory.research_hub import detail_page as detail_page_factory
 from networkapi.wagtailpages.tests.research_hub import base as research_test_base
 from networkapi.wagtailpages.tests.research_hub import utils as research_test_utils
 
@@ -14,7 +15,7 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
     @classmethod
     def setUpTestData(cls):
         super().setUpTestData()
-        cls.detail_page = research_factory.ResearchDetailPageFactory(
+        cls.detail_page = detail_page_factory.ResearchDetailPageFactory(
             parent=cls.library_page,
             original_publication_date=(research_test_utils.days_ago(n=14)),
         )
@@ -134,7 +135,7 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
             self.detail_page,
             self.fr_locale,
         )
-        extra_detail_page = research_factory.ResearchDetailPageFactory(
+        extra_detail_page = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
         )
         research_factory.ResearchAuthorRelationFactory(
@@ -167,7 +168,7 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
             self.detail_page,
             self.fr_locale,
         )
-        extra_detail_page = research_factory.ResearchDetailPageFactory(
+        extra_detail_page = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
         )
         research_factory.ResearchAuthorRelationFactory(
@@ -195,15 +196,15 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
         self.assertIn(fr_extra_detail_page, context["latest_research"])
 
     def test_get_latest_research(self):
-        detail_page_1 = research_factory.ResearchDetailPageFactory(
+        detail_page_1 = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
             original_publication_date=(research_test_utils.days_ago(n=3)),
         )
-        detail_page_2 = research_factory.ResearchDetailPageFactory(
+        detail_page_2 = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
             original_publication_date=(research_test_utils.days_ago(n=2)),
         )
-        detail_page_3 = research_factory.ResearchDetailPageFactory(
+        detail_page_3 = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
             original_publication_date=(research_test_utils.days_ago(n=1)),
         )
